Rallypoint Changes, Start of Sprites being implemented
Rallypoints now contain additional data so spell logic can be done.

Buildings and Units (soon resources and other entitiescan now have
sprites ).

Unit sprites are split into folders:
Idle - For when units are not moving
IdleWithResources - For when units are not moving but holding resources
(trees, apples, minerals, vespene gas, etc)
Moving - When units are moving
MovingWithResources - Same as before, but with resources, etc etc

Buildings have folders for their types
Standard - When the building is doing nothing
Producing - When the building is making a unit
BeingBuild - NOT TO BE CONFUSED WITH PRODUCING, this is shown when the
actual building is being constructed.

Currently, the sprite animation sequence is deterimined alphebetically
(spelled that wrong probably), so you'll notice most folders the sprites
are numbered 1-9.png, and goes into AA,BB,CC, etc. This will most likely
be changed into something more user friendly.

-CEO of Ausware.